% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.bgm_prior_ess}
\alias{summary.bgm_prior_ess}
\title{Summary method for prior_ess objects}
\usage{
\method{summary}{bgm_prior_ess}(object, ...)
}
\description{
Summary method for prior_ess objects
}
