% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designbgm-package.R
\docType{package}
\name{designbgm-package}
\alias{designbgm}
\alias{designbgm-package}
\title{designbgm: Sample Size Planning for Bayesian Graphical Models}
\description{
Plans sample sizes of a prospective study for a Bayesian graphical model, given a prior elicited from a previous study. Two planning methods are available: a data-to-prior information ratio over the model parameters, and Bayes factor design analysis at a representative edge. Estimators of the prior effective sample size quantify how much information the elicited prior carries, and planned sample sizes can be validated by simulation. Gaussian graphical models are currently supported.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Bayesian-Graphical-Modelling-Lab/designbgm}
  \item Report bugs at \url{https://github.com/Bayesian-Graphical-Modelling-Lab/designbgm/issues}
}

}
\author{
\strong{Maintainer}: Giuseppe Arena \email{g.arena@uva.nl} (\href{https://orcid.org/0000-0001-5204-3326}{ORCID})

}
\keyword{internal}
